WIP: Working towards major pipeline cleanup.
Two majors efforts in progress:

1. Removal of priority based streaming and fast path codes in
preparation for future refactoring. These were polluting core
classes despite having a small user base. In the future, they
will be refactored into the appropriate subclasses and "plugins".

2. Refactoring the way VTK goes between piece and structured
extents. Before, extent translators were used when the pipeline
moved from structured to unstructured data converting piece
request to extent request. This caused many problems with filters
that altered extents, mainly a lot of redundant IO due to
repartitioning of different extents. This became extremely
cumbersome to manage when running distributed. The new behavior
pushes the extent translation all to way to the readers and
only when readers are able to read a subset. This works much
better. The only downside is that filters need to be able to
handle data extents different than update extents. Most filters
can do this but many imaging filters cannot. Those that are
needed in parallel will have to be updated.